requirements.txt
czifile~=2019.7.2
numpy~=1.20.3
openpyxl >= 3.0.0
pandas~=1.2.5
Pillow~=8.3.1
pysimplegui==4.45.0
python-bioformats==4.0.4
PyYAML~=5.4.1
scikit-image~=0.18.2
Shapely~=1.7.1
tifffile~=2021.7.2
